// guipmc-casuistica/src/app/casuistica/models/auditorias.ts
import { CamposOriginalesDTO } from './camposOriginalesDTO';
export class Auditorias{
  nomUsuario: string;
  numFolioMovOriginal: string;
  desObservacionesAprobador: string;
  desObservacionesSol: string;
  cveIdAccionRegistro: number;
  desAccionRegistro: string;
  desCambio: string;
  desSituacionRegistro: string;
  camposOriginalesDTO: CamposOriginalesDTO;
  fecAlta:string;
  fecBaja:string;

}